Python HackerRank Challenges

+++ Python/Numpy/zeros_ones.py
@@ -3,10 +3,9 @@ import numpy as np
 
 def main() -> None:
     dims = tuple(map(int, input().split()))
-    print(np.zeros(dims, dtype=np.int32))
-    print(np.ones(dims, dtype=np.int32))
+    print(np.zeros(dims, dtype=np.int))
+    print(np.ones(dims, dtype=np.int))
 
 
 if __name__ == "__main__":
     main()
-
